+++ test/LayeredCraft.DecoWeaver.Generator.Tests/Cases/024_ServiceDecorator_MultipleOrdered/Global.cs
@@ -1,0 +1,4 @@
+using DecoWeaver.Sample;
+
+[assembly: DecoWeaver.Attributes.DecorateService(typeof(IRepository<>), typeof(CachingRepository<>), 1)]
+[assembly: DecoWeaver.Attributes.DecorateService(typeof(IRepository<>), typeof(LoggingRepository<>), 2)]