+++ test/LayeredCraft.DecoWeaver.Generator.Tests/Cases/026_SkipAssemblyWithClassLevel/Program.cs
@@ -1,0 +1,8 @@
+using Microsoft.Extensions.DependencyInjection;
+
+var services = new ServiceCollection();
+
+// Register SpecialRepository - should get ValidationRepository but NOT CachingRepository
+services.AddScoped<IRepository<User>, SpecialRepository<User>>();
+
+public class User { }