+++ test/LayeredCraft.DecoWeaver.Generator.Tests/Cases/028_MergePrecedence_SortOrder/Global.cs
@@ -1,0 +1,4 @@
+using DecoWeaver.Sample;
+
+// Assembly-level: Logging@10 for IRepository<>
+[assembly: DecoWeaver.Attributes.DecorateService(typeof(IRepository<>), typeof(LoggingRepository<>), 10)]