+++ test/LayeredCraft.DecoWeaver.Generator.Tests/Cases/030_DoNotDecorate_Multiple/Global.cs
@@ -1,0 +1,6 @@
+using DecoWeaver.Sample;
+
+// Assembly declares 3 decorators for all IRepository<> implementations
+[assembly: DecoWeaver.Attributes.DecorateService(typeof(IRepository<>), typeof(CachingRepository<>), 10)]
+[assembly: DecoWeaver.Attributes.DecorateService(typeof(IRepository<>), typeof(LoggingRepository<>), 20)]
+[assembly: DecoWeaver.Attributes.DecorateService(typeof(IRepository<>), typeof(ValidationRepository<>), 30)]